[REVERT ME]: mmgr : Do not open ttys in blocking mode
BZ: 71590

This patch removes the O_NONBLOCK Flag when opening ttyIFX0
and then the open of ttyIFX0 is done in blocking mode. the return
will happens only when the device is ready to process command and
read/write data. This patch reduce the occurence of the issue
in BZ 71590.

Signed-off-by: alain kouassu <alain.kouassu@intel.com>

This patch apply on top of manifest-20130128_1305-generated.xml

Change-Id: I286f6e6b86a9b884b32a5ffdcd0f36907aa98e51
Reviewed-on: http://android.intel.com:8080/88975
Reviewed-by: Lucas, Guillaume <guillaume.lucas@intel.com>
Reviewed-by: Robert, Denis <denis.robert@intel.com>
Tested-by: Robert, Denis <denis.robert@intel.com>
Reviewed-by: De Oliveira, CesarX <cesarx.de.oliveira@intel.com>
Reviewed-by: Ravindran, Arun <arun.ravindran@intel.com>
Reviewed-by: buildbot <buildbot@intel.com>
Tested-by: buildbot <buildbot@intel.com>