mmgr: shutdown modem when no client connected.
BZ: 82289

When a client is disconnected from mmgr, the associated modem resource must be unlocked. And
if it is the last connected client modem must be shutdowned.

At client disconnection handling modem resource is unlocked and if the modem resource is
not locked by any mmgr client(s) then a shutdown request is sent to modem.

Change-Id: I8f5b2d6bee06e3d5223ebefc2c29ea837538e91b
Signed-off-by: David Bagnol <david.bagnol@intel.com>
Reviewed-on: http://android.intel.com:8080/99491
Reviewed-by: cactus <cactus@intel.com>
Reviewed-by: Lucas, Guillaume <guillaume.lucas@intel.com>
Reviewed-by: De Oliveira, CesarX <cesarx.de.oliveira@intel.com>
Reviewed-by: Ravindran, Arun <arun.ravindran@intel.com>
Reviewed-by: Lebsir, SamiX <samix.lebsir@intel.com>
Tested-by: Lebsir, SamiX <samix.lebsir@intel.com>
Reviewed-by: buildbot <buildbot@intel.com>
Tested-by: buildbot <buildbot@intel.com>