[REVERT ME] mmgr: never power on the modem if telephony stack is disabled
BZ: 99081

To enable the platform power management, the telephony stack should be
disabled as we are currently facing a HSIC issue.

For that purpose:
* AMTL should be updated to be able to turn on or off the telephony stack.
A platform reboot will be needed to take this modification into account. An
Android property will be set according to this configuration.
* MMGR should be updated to handle the Android property set by AMTL. If the
telephony stack is disabled, MMGR will inform its clients that the modem is out
of service. The modem remains off.

This feature is only enabled for saltbay products

Change-Id: I6c6cb881c20a75292febd4ebb007a4dc7dc64f5b
Signed-off-by: Cesar DE OLIVEIRA <cesarx.de.oliveira@intel.com>
Reviewed-on: http://android.intel.com:8080/100742
Reviewed-by: Rodrigues, JoseX <josex.rodrigues@intel.com>
Reviewed-by: Butscher, MorganeX <morganex.butscher@intel.com>
Reviewed-by: cactus <cactus@intel.com>
Reviewed-by: Lucas, Guillaume <guillaume.lucas@intel.com>
Reviewed-by: Ravindran, Arun <arun.ravindran@intel.com>
Tested-by: Louis, NicolasX <nicolasx.louis@intel.com>
Reviewed-by: buildbot <buildbot@intel.com>
Tested-by: buildbot <buildbot@intel.com>