[PORT FROM R42S] mmgr: change the wait loop after MUX activation
BZ: 93966

As device names are created in devfs as soon as the GSMIOC_GETCONF
ioctl is called, using the presence of /dev/gsmtty63 to check the MUX
readiness is not correct: this may lead to EL2NSYNC errors as clients
can open a TTY while channel 0 has not been open yet in the kernel
(scenario that is not supported by the latter).

This patch changes the 'stat' check by actually trying to open a
specified TTY (by default /dev/gsmtty1).

Orig-Change-Id: Iecb1bae6329184dfc7a2c3b3a25d3b078d097012
Change-Id: Ife30d62b765e7a0d98533f32436693e365397518
Signed-off-by: Lionel Ulmer <lionel.ulmer@intel.com>
Signed-off-by: Edward Marmounier <edward.marmounier@intel.com>
Reviewed-on: http://android.intel.com:8080/102067
Reviewed-by: De Oliveira, CesarX <cesarx.de.oliveira@intel.com>
Reviewed-by: cactus <cactus@intel.com>
Reviewed-by: buildbot <buildbot@intel.com>
Tested-by: buildbot <buildbot@intel.com>