mmgr: add IPC power management
BZ: 90679

IPC power management was done in MCDR. This code was not agnostic. It
only handles the UART link.

This code has been moved in MMGR because it's more relevant to have it
in this component. MMGR needs to configure the IPC power management for
core dump retrieval and flashing purpose.

Change-Id: I88cefc71412a6c1b2f1cac9ea450bce8e5b3dc5f
Signed-off-by: Cesar DE OLIVEIRA <cesarx.de.oliveira@intel.com>
Reviewed-on: http://android.intel.com:8080/101614
Reviewed-by: cactus <cactus@intel.com>
Reviewed-by: Bagnol, David <david.bagnol@intel.com>
Reviewed-by: Ravindran, Arun <arun.ravindran@intel.com>
Tested-by: Louis, NicolasX <nicolasx.louis@intel.com>
Reviewed-by: buildbot <buildbot@intel.com>
Tested-by: buildbot <buildbot@intel.com>