[PORT FROM R42S] mmgr: make the modem_shutdown process compliant to the SAS
BZ: 99482

As per CLVT SAS, MODEM_DOWN event must be sent after doing
the modem power down.

But in code it is done before. So modem is not yet down.
And it takes several seconds to do that.
So AP can be shutdown before having the time to close properly the modem.

Operations to be done in MMgr modem_shutdwn():
1/ state = MDM_CTRL_STATE_OFF
2/ send AT+CFUN=0
3/ wait for result (whatever OK or ERROR)
4/ close DLC for at cmd
5/ == send MODEM_DOWN notification ==
6/ close ifx0
7/ close modem

(MODEM_SHUTDOWN is supposed to have been sent before
modem_down() to notify the clients)

Change-Id: I7b6bb97c597d92eb167478727f23bd760e305ff6
Orig-Change-Id: Iacf944832109e0fe3476ca29ed589befbd98f4ec
Signed-off-by: Morgane BUTSCHER <morganex.butscher@intel.com>
Reviewed-on: http://android.intel.com:8080/108743
Reviewed-by: cactus <cactus@intel.com>
Reviewed-by: Ravindran, Arun <arun.ravindran@intel.com>
Reviewed-by: Badrappan, Jeevaka <jeevaka.badrappan@intel.com>
Reviewed-by: Louis, NicolasX <nicolasx.louis@intel.com>
Tested-by: Louis, NicolasX <nicolasx.louis@intel.com>
Reviewed-by: buildbot <buildbot@intel.com>
Tested-by: buildbot <buildbot@intel.com>