mmgr: add reconnect feature to libmmgrcli
BZ: 101744

Clients are not aware of their disconnection after a MGMR crash for
example. Thus, they are not able to reestablish the connection.

The libmmgrcli hides the disconnection. When it does happen, the library
notifies a fake modem down event by calling the client's callback function.
Then, it reconnects to the mmgr and restores the client context (it acquires the
resource if the client already holds it).

Change-Id: Ibfd52f5a8154866652065456ae5690b629c2d73c
Signed-off-by: Cesar DE OLIVEIRA <cesarx.de.oliveira@intel.com>
Reviewed-on: http://android.intel.com:8080/104805
Reviewed-by: cactus <cactus@intel.com>
Reviewed-by: Butscher, MorganeX <morganex.butscher@intel.com>
Reviewed-by: Ulmer, Lionel <lionel.ulmer@intel.com>
Reviewed-by: Rodrigues, JoseX <josex.rodrigues@intel.com>
Reviewed-by: Ravindran, Arun <arun.ravindran@intel.com>
Reviewed-by: Louis, NicolasX <nicolasx.louis@intel.com>
Tested-by: Louis, NicolasX <nicolasx.louis@intel.com>
Reviewed-by: buildbot <buildbot@intel.com>
Tested-by: buildbot <buildbot@intel.com>