[PORT FROM R42S] mmgr: add the core dump event in the MMgr poll list for resource acquire
BZ: 94563

When the MMgr wakes up the modem after a resource acquire, the core
dump event is not in the poll list. So if a core dump is raised when
the MMgr is waiting the IPC_READY even, this one will not be take
into account and the IPC_READY timeout will be raised.

This patch fixes the MMgr code by adding the core dump event in the
poll list for the resource acquire operation.

Change-Id: Iddc848d117a618e0fef529358416dd8ff1c58db5
Signed-off-by: Guillaume Lucas <guillaume.lucas@intel.com>
Signed-off-by: Cesar DE OLIVEIRA <cesarx.de.oliveira@intel.com>
Reviewed-on: http://android.intel.com:8080/102870
Reviewed-by: cactus <cactus@intel.com>
Reviewed-by: Butscher, MorganeX <morganex.butscher@intel.com>
Reviewed-by: Ravindran, Arun <arun.ravindran@intel.com>
Tested-by: Louis, NicolasX <nicolasx.louis@intel.com>
Reviewed-by: buildbot <buildbot@intel.com>
Tested-by: buildbot <buildbot@intel.com>