[PORT FROM R42S] mmgr: fix file descriptor closure
BZ: 97929

In the reset_modem function fd_tty is not closed when
an operation is skipped.

Moreover if a file descriptor is not correctly closed, when an event happened
MMGR will interpret it as a client event. But MMGR will fail to find the right
client

open_tty should not overwrite an existing fd.
close the fd when the client is not found

Change-Id: I4c6c404fdfaf692dff340dc166546092fe77470d
Signed-off-by: Morgane BUTSCHER <morganex.butscher@intel.com>
Signed-off-by: Cesar DE OLIVEIRA <cesarx.de.oliveira@intel.com>
Reviewed-on: http://android.intel.com:8080/101913
Reviewed-by: Marmounier, Edward <edward.marmounier@intel.com>
Reviewed-by: Badrappan, Jeevaka <jeevaka.badrappan@intel.com>
Reviewed-by: Ravindran, Arun <arun.ravindran@intel.com>
Reviewed-by: Rodrigues, JoseX <josex.rodrigues@intel.com>
Tested-by: Louis, NicolasX <nicolasx.louis@intel.com>
Reviewed-by: buildbot <buildbot@intel.com>
Tested-by: buildbot <buildbot@intel.com>