mmgr: set C client library API synchronous
BZ: 108312

When MMGR is busy, clients' requests are rejected. In that case, MMGR
sends a NACK answer. Users were asked to handle themselves the retry
mechanism. Unfortunately, most of them don't handle it. In addition,
there is also a bug in the client library. If the resource request is
rejected, the client is not able to resend the request. This may lead to
several issues. Worst case: RIL is unable to get the resource. As a
consequence, the modem remains off.

The C client library API is now synchronous. The library subscribes
itself to ACK/NACK events and wait for those acknowledge messages.
A timeout of 20s is used to handle MMGR's unresponsiveness.
MMGR code has been updated to send acknowledge messages ASAP.

Change-Id: I195560b0f99b78a2e5c0abfe817044f3dd005e58
Signed-off-by: Cesar DE OLIVEIRA <cesarx.de.oliveira@intel.com>
Reviewed-on: http://android.intel.com:8080/107759
Reviewed-by: cactus <cactus@intel.com>
Reviewed-by: Ulmer, Lionel <lionel.ulmer@intel.com>
Reviewed-by: Ravindran, Arun <arun.ravindran@intel.com>
Reviewed-by: Rodrigues, JoseX <josex.rodrigues@intel.com>
Reviewed-by: Butscher, MorganeX <morganex.butscher@intel.com>
Reviewed-by: Louis, NicolasX <nicolasx.louis@intel.com>
Tested-by: Louis, NicolasX <nicolasx.louis@intel.com>
Reviewed-by: buildbot <buildbot@intel.com>
Tested-by: buildbot <buildbot@intel.com>