mmgr-test: reports test failed when a bad reception occured
BZ: 114390

A test should be declared FAILED if the received data from MMGR
is empty while it should not be.

mmgr-test has an event structure which indicates if a reception
error occurs.  The test is declared UNCONCLUSIVE when MMGR
notifies a modem OOS or platform reboot event, except for the
full escalation recovery test.

Change-Id: I5311c689e7b3d04425ca8a9cc6bce71676e1faca
Signed-off-by: Cesar DE OLIVEIRA <cesarx.de.oliveira@intel.com>
Reviewed-on: http://android.intel.com:8080/112274
Reviewed-by: cactus <cactus@intel.com>
Reviewed-by: Badrappan, Jeevaka <jeevaka.badrappan@intel.com>
Reviewed-by: Akue, LoicX <loicx.akue@intel.com>
Reviewed-by: Lebsir, SamiX <samix.lebsir@intel.com>
Tested-by: Lebsir, SamiX <samix.lebsir@intel.com>
Reviewed-by: buildbot <buildbot@intel.com>
Tested-by: buildbot <buildbot@intel.com>