mmgr: secure channel - encoding/decoding messages
BZ: 103694

Modem sends and receives messages encoded in base64 format via
its secured channel dedicated DLC. MMGR should decode the
message sent by the modem, give it to the security library,
encode the response of the security library and send it back to
the modem.

This feature is still disabled by default.

Change-Id: Idd72dc67db78d1dcb51127fe89992c2ebc2af57e
Signed-off-by: Cesar DE OLIVEIRA <cesarx.de.oliveira@intel.com>
Reviewed-on: http://android.intel.com:8080/104807
Reviewed-by: cactus <cactus@intel.com>
Reviewed-by: Mcgillion, Brian <brian.mcgillion@intel.com>
Reviewed-by: Badrappan, Jeevaka <jeevaka.badrappan@intel.com>
Reviewed-by: Lebsir, SamiX <samix.lebsir@intel.com>
Tested-by: Lebsir, SamiX <samix.lebsir@intel.com>
Reviewed-by: buildbot <buildbot@intel.com>
Tested-by: buildbot <buildbot@intel.com>