libmmgr: fix reconnect feature
BZ: 117076

Since patch 107759, the client library is synchronous and the
send_msg function waits for MMGR's answer from the reader thread.
When the client is disconnected, the reader thread detects the
disconnection and tries to reconnect the client. The function
send_msg is used for that purpose. Obviously, no answer will be
received from MMGR as the reader thread is already busy...

send_msg function is now able to read itself MMGR's answer when
needed. The client context is also correctly restored now.
The client registration is also limited by a 20s timeout.

Change-Id: Ic99c0662bef8d749e3c061069e85f4089d1ddbc2
Signed-off-by: Cesar DE OLIVEIRA <cesarx.de.oliveira@intel.com>
Reviewed-on: http://android.intel.com:8080/115404
Reviewed-by: Ravindran, Arun <arun.ravindran@intel.com>
Tested-by: Louis, NicolasX <nicolasx.louis@intel.com>
Reviewed-by: buildbot <buildbot@intel.com>
Tested-by: buildbot <buildbot@intel.com>