libmmgrcli: fix API description
BZ: 115999

A deadlock of 20s happens when a client tries to send a message
under their callback. The contract between the lib and the client
is not as clear as it should be.

Client library modifications:
- better API description. Highlighted points:
  - timeouts have been detailed.
  - client's callback must be used only for a short processing
    time, otherwise we do not guarantee the responsiveness.
- Clients are not allowed to send a message to MMGR under their
callback.
- Clients are no more requested to explicitly send ACK_COLD_RESET
or ACK_MODEM_SHUTDOWN. This message is sent by the client
library according to the returned value of the callback.
- Registration to COLD RESET or MODEM SHUTDOWN is no more
mandatory. Only clients which have specific operation to do on
this event should subscribe.

MMGR modifications:
- ACK_COLD_RESET and ACK_MODEM_SHUTDOWN requests are no more
ACKed by MMGR.
- If the client is not registered to an event, MMGR will not
wait for its ACK.

mmgr-test:
removal of ACK_COLD_RESET and ACK_MODEM_SHUTDOWN

General:
removal of outdated README.txt file.

Change-Id: Iae6832ea16c093a8166117c18bc2f809c022b9e9
Signed-off-by: Cesar DE OLIVEIRA <cesarx.de.oliveira@intel.com>
Reviewed-on: http://android.intel.com:8080/113511
Reviewed-by: Ravindran, Arun <arun.ravindran@intel.com>
Reviewed-by: Lucas, Guillaume <guillaume.lucas@intel.com>
Tested-by: Louis, NicolasX <nicolasx.louis@intel.com>
Reviewed-by: buildbot <buildbot@intel.com>
Tested-by: buildbot <buildbot@intel.com>