[REVERT ME] mmgr: NVM backup
BZ: 115634

If NVM files are corrupted, flashless modem will never boot.
Currently, if the flash failed, we have no way of knowing if
this is due to NVM corruption. But, most of the time, it is the
case.

A new component will allow MMGR to check the consistency of the
NVM files, but in the meantime, this W/A will be used:
- if MMGR has received IPC_READY, NVM files will be backup.
- if the flash operation has failed, NVM files will be restored.
  Blank files will be used if no backup files.

Change-Id: I5337a6511693c793fb28f5573e0fcca35b9a46e0
Signed-off-by: Cesar DE OLIVEIRA <cesarx.de.oliveira@intel.com>
Reviewed-on: http://android.intel.com:8080/118439
Reviewed-by: Rodrigues, JoseX <josex.rodrigues@intel.com>
Reviewed-by: Ulmer, Lionel <lionel.ulmer@intel.com>
Reviewed-by: Lucas, Guillaume <guillaume.lucas@intel.com>
Reviewed-by: Ravindran, Arun <arun.ravindran@intel.com>
Tested-by: Louis, NicolasX <nicolasx.louis@intel.com>
Reviewed-by: buildbot <buildbot@intel.com>
Tested-by: buildbot <buildbot@intel.com>