libmgmrcli: fix timeout when clients explicitly send ACK
BZ: 122608

When a client use the mmgr_cli_send_msg function to send the
ACK_COLD_RESET or ACK_MDM_SHUTDOWN, the client will face a
deadlock of 20s. Thus, because the client library is waiting for
an acknowledge.

When those requests are sent, do not wait for a ACK. This ACK
message is not sent by MMGR for those kind of messages.

Change-Id: I6f1165a6c85e53e493503d9e0ec1bcbeebc89b41
Signed-off-by: Cesar DE OLIVEIRA <cesarx.de.oliveira@intel.com>
Reviewed-on: http://android.intel.com:8080/119718
Reviewed-by: Ulmer, Lionel <lionel.ulmer@intel.com>
Reviewed-by: Goea, Andrei DanielX <andrei.danielx.goea@intel.com>
Reviewed-by: Roulliere, Pierre <pierre.roulliere@intel.com>
Reviewed-by: Lucas, Guillaume <guillaume.lucas@intel.com>
Reviewed-by: Marmounier, Edward <edward.marmounier@intel.com>
Reviewed-by: Badrappan, Jeevaka <jeevaka.badrappan@intel.com>
Reviewed-by: Ravindran, Arun <arun.ravindran@intel.com>
Tested-by: Louis, NicolasX <nicolasx.louis@intel.com>
Reviewed-by: buildbot <buildbot@intel.com>
Tested-by: buildbot <buildbot@intel.com>