mmgr: core dump notification enhancement
BZ: 124388

If MMGR fails to turn on the modem after a core dump detection,
the core dump is never notified as MMGR needs to extract the
panic ID from the core dump for it to be signaled.

With the new design, the panic ID is not required anymore,
therefore its extraction is removed from the code. The core dump
is notified once the retrieval operation is completed.
To improve crashltool statistics, additional information is
provided.

Change-Id: I3089b61f069da22167e9e2ac4a31e05fc3a33f30
Signed-off-by: Cesar DE OLIVEIRA <cesarx.de.oliveira@intel.com>
Reviewed-on: http://android.intel.com:8080/119315
Reviewed-by: Roulliere, Pierre <pierre.roulliere@intel.com>
Reviewed-by: Rodrigues, JoseX <josex.rodrigues@intel.com>
Reviewed-by: Ulmer, Lionel <lionel.ulmer@intel.com>
Reviewed-by: Goea, Andrei DanielX <andrei.danielx.goea@intel.com>
Reviewed-by: Ravindran, Arun <arun.ravindran@intel.com>
Reviewed-by: Lucas, Guillaume <guillaume.lucas@intel.com>
Tested-by: Louis, NicolasX <nicolasx.louis@intel.com>
Reviewed-by: cactus <cactus@intel.com>
Tested-by: cactus <cactus@intel.com>