mmgr: stop client thred if connectAsync failed
BZ: 124227

If the connectAsync() operation failed due to an ACK timeout,
the MMGR client thread is not stoped. Due to this, the next
call to connectAsync() returns immediatly even if the
connection is not fully performed due to the previous ACK
timeout.

This patch call the stop() method if we received an ACK timeout
during the connectAsync() operation.

Change-Id: I0c3ee8c54d22781a1a89f8e672bcb51ee6f4a24b
Signed-off-by: Guillaume Lucas <guillaume.lucas@intel.com>
Reviewed-on: http://android.intel.com:8080/120134
Reviewed-by: Marmounier, Edward <edward.marmounier@intel.com>
Reviewed-by: Ulmer, Lionel <lionel.ulmer@intel.com>
Reviewed-by: De Oliveira, CesarX <cesarx.de.oliveira@intel.com>
Reviewed-by: Ravindran, Arun <arun.ravindran@intel.com>
Tested-by: Louis, NicolasX <nicolasx.louis@intel.com>
Reviewed-by: buildbot <buildbot@intel.com>
Tested-by: buildbot <buildbot@intel.com>