Platform does not enter s0ix state
BZ: 124698

Platform cannot enter s0ix state due a side effect in library
loading (dlerror not checked).

On some functions, we are trying to load a library, that might
be missing. In this case, dlerror will be set and can lead to
an error on the next library load. Solution is to get dlerror
after a call to dlopen before to leave the function.

It's already what is made in other parts of the code.

Change-Id: Ia4ade017f941708599c75b2ecc847461a93a5f47
Signed-off-by: Eric Arnould <ericx.arnould@intel.com>
Reviewed-on: http://android.intel.com:8080/120741
Reviewed-by: Muraru, PetruX <petrux.muraru@intel.com>
Reviewed-by: Garipoli, Giuseppe <giuseppe.garipoli@intel.com>
Reviewed-by: Lucas, Guillaume <guillaume.lucas@intel.com>
Reviewed-by: Le Meur, ArnaudX <arnaudx.le.meur@intel.com>
Reviewed-by: De Oliveira, CesarX <cesarx.de.oliveira@intel.com>
Reviewed-by: Predon, Frederic <frederic.predon@intel.com>
Reviewed-by: Blagny, ChristianX <christianx.blagny@intel.com>
Tested-by: Louis, NicolasX <nicolasx.louis@intel.com>
Reviewed-by: buildbot <buildbot@intel.com>
Tested-by: buildbot <buildbot@intel.com>