[REVERTME] mmgr: increase core dump retrieve timeout
BZ: 128384

For CTP7160, Core dump retrieval takes more time than timeout
configured in mmgr and mmgr-test. Then the core dump cannot
be stored on the file system.
This patch increase the timeout to be able to upload complete
core dumps to crashtool.
It should be reverted as soon as a solution to reduce retrieval
in less than 10mn is available (YModem, compression,...)

Change-Id: I3d7d00f33b6837253792438cb2eaffff0b493878
Signed-off-by: Le Meur, ArnaudX <arnaudx.le.meur@intel.com>
Reviewed-on: http://android.intel.com:8080/122962
Reviewed-by: buildbot <buildbot@intel.com>
Reviewed-by: Blagny, ChristianX <christianx.blagny@intel.com>
Reviewed-by: De Oliveira, CesarX <cesarx.de.oliveira@intel.com>
Reviewed-by: Ravindran, Arun <arun.ravindran@intel.com>
Reviewed-by: Lebsir, SamiX <samix.lebsir@intel.com>
Tested-by: Lebsir, SamiX <samix.lebsir@intel.com>
Tested-by: Louis, NicolasX <nicolasx.louis@intel.com>
Reviewed-by: cactus <cactus@intel.com>
Tested-by: cactus <cactus@intel.com>