mmgr: warm reset operation removal
BZ: 130546

By experience, we know that the warm reset operation is not
efficient. This warm reset operation is the less efficient
operation and the most complex to support. In several cases,
this operation is skipped (after a modem self-reset, succeed
core dump retrieval, etc). A useless context needs also to be
handled for this.

Let’s be smart and get rid of useless specification and code.

Change-Id: I64d69cfaa003e763bf330c1904c9ad2e8dcedc61
Signed-off-by: Cesar DE OLIVEIRA <cesarx.de.oliveira@intel.com>
Reviewed-on: http://android.intel.com:8080/125311
Reviewed-by: Marmounier, Edward <edward.marmounier@intel.com>
Reviewed-by: Roulliere, Pierre <pierre.roulliere@intel.com>
Reviewed-by: Ravindran, Arun <arun.ravindran@intel.com>
Reviewed-by: Bagnol, David <david.bagnol@intel.com>
Tested-by: Louis, NicolasX <nicolasx.louis@intel.com>
Reviewed-by: cactus <cactus@intel.com>
Tested-by: cactus <cactus@intel.com>