mmgr: Add timeout for modem enumeration in core dump mode
BZ: 132135

If the modem never re-enumerates in core dump mode after signaling
the core dump to AP, mmgr will remain blocked waiting for the event
leading to infinite unavailability of the modem.
Adding a timeout of 10 seconds between the modem the core dump signal
is received and when the modem core dump is being retreived.

Change-Id: Ie742f8b64c824f10af219a13d312cb36717c8ed8
Signed-off-by: Goea, Andrei DanielX <andrei.danielx.goea@intel.com>
Reviewed-on: http://android.intel.com:8080/127645
Reviewed-by: Louis, NicolasX <nicolasx.louis@intel.com>
Reviewed-by: Ravindran, Arun <arun.ravindran@intel.com>
Reviewed-by: Lebsir, SamiX <samix.lebsir@intel.com>
Tested-by: Lebsir, SamiX <samix.lebsir@intel.com>
Reviewed-by: cactus <cactus@intel.com>
Tested-by: cactus <cactus@intel.com>