[MMGR]: Reset CORE DUMP HSIC link after a delay
BZ: 136383

For Baytrail platform MMGR must reset the HSIC link in
case of core dump after 6 seconds, because of an HW issue.
The patch uses a new parameter from MMGR.CONF file. This
parameter sets a timer after the core dump detection.
When this timer expires, the link used for the core
dump is reset. If this paramter is 0, no link reset
is done.

Change-Id: I2ed72fdf61dd467ce08f31f923361df4ff8c0be5
Signed-off-by: Gwenn Bourree <gwenn.bourree@intel.com>
Reviewed-on: http://android.intel.com:8080/132219
Reviewed-by: Delmas-begue, Ulysse <ulysse.delmas-begue@intel.com>
Reviewed-by: Lucas, Guillaume <guillaume.lucas@intel.com>
Reviewed-by: De Oliveira, CesarX <cesarx.de.oliveira@intel.com>
Reviewed-by: Ravindran, Arun <arun.ravindran@intel.com>
Tested-by: Louis, NicolasX <nicolasx.louis@intel.com>
Reviewed-by: cactus <cactus@intel.com>
Tested-by: cactus <cactus@intel.com>