[REVERT ME][MMGR]: disable HSIC PM when core dump complete
BZ: 139075

In order to ensure the modem enumeration will be
detected by the AP HSIC driver when the modem self
reset after the core dump evacuation, the MMGR must
disable the HSIC power management. Because the HSIC
power management is already disabled during the core
dump  file evacuation, there is then nothing to do.
The patch remove the HSIC PM enabling when core dump
is complete. This patch must be reverted when HSIC
Power management is completely validated: Indeed, the
AUX interrupt raised by the modem during its self-reset
has to disable the power management that ensure the
detection of the HSIC enumeration.

Change-Id: I041ed6a0a9ff5a904e8e36b5cbc92de9b773aa12
Signed-off-by: Gwenn Bourree <gwenn.bourree@intel.com>
Reviewed-on: http://android.intel.com:8080/132587
Reviewed-by: Tancerel, LudovicX <ludovicx.tancerel@intel.com>
Reviewed-by: Delmas-begue, Ulysse <ulysse.delmas-begue@intel.com>
Reviewed-by: Lucas, Guillaume <guillaume.lucas@intel.com>
Reviewed-by: Ulmer, Lionel <lionel.ulmer@intel.com>
Reviewed-by: Goea, Andrei DanielX <andrei.danielx.goea@intel.com>
Reviewed-by: De Oliveira, CesarX <cesarx.de.oliveira@intel.com>
Reviewed-by: Ravindran, Arun <arun.ravindran@intel.com>
Tested-by: Louis, NicolasX <nicolasx.louis@intel.com>
Reviewed-by: cactus <cactus@intel.com>
Tested-by: cactus <cactus@intel.com>