mmgr: log enhancement
BZ: 141680

MMGR and libmmgrcli logs are sometimes mixed up. This is due a
LOG_TAG conflict.

LOG_TAG is now set with a CFLAG set in Android.mk (one different
per component).
Logs macros for libmmgrcli have been updated to always print
the name and the file descriptor of the client.

Change-Id: I499102512e3695e14993fdcddc4e14670b61ede4
Signed-off-by: Cesar DE OLIVEIRA <cesarx.de.oliveira@intel.com>
Reviewed-on: http://android.intel.com:8080/134894
Reviewed-by: Ulmer, Lionel <lionel.ulmer@intel.com>
Reviewed-by: Rodrigues, JoseX <josex.rodrigues@intel.com>
Reviewed-by: Roulliere, Pierre <pierre.roulliere@intel.com>
Reviewed-by: Lucas, Guillaume <guillaume.lucas@intel.com>
Tested-by: Louis, NicolasX <nicolasx.louis@intel.com>
Reviewed-by: Louis, NicolasX <nicolasx.louis@intel.com>
Reviewed-by: Ravindran, Arun <arun.ravindran@intel.com>
Reviewed-by: cactus <cactus@intel.com>
Tested-by: cactus <cactus@intel.com>