mmgr: bad firmware family
BZ: 144594

In case of E_MODEM_FW_BAD_FAMILY, modem is not properly
shutdown.

The correct way to turn off the modem and declare it OOS is to
set MMGR state to MDM_RESET to call the recovery module and
force modem recovery to OOS. modem_shutdown should not be called
explicitly.

Change-Id: I22b2b2c6b5b11e8abac67a51f1a36eeb18c6d662
Signed-off-by: Cesar DE OLIVEIRA <cesarx.de.oliveira@intel.com>