mmgr: handle modem acquisition in modem configuration phase
BZ: 146644

If a second client tries to acquire the modem while it is being
configured (power up / flash / MUX / ...) MMGR NACKs the request
leading to this acquire to be periodically retried.

This introduces delay in the handling of RADIO_POWER in the RIL
code if another client has acquired the modem first. This leads
to bad user experience in cases like emergency call while in
airplane mode.

The fix is to handle the acquire calls during the MDM_START and
MDM_CONF_ON_GOING states.

Change-Id: Ie311f9f5002120d0e26f5ba18630300f05a8f736
Signed-off-by: Lionel Ulmer <lionel.ulmer@intel.com>