[Telephony/MMGR]: Add 7260_CONF_3 and usb ipc mgt
BZ: 144265

In order to support the SSIC IPC link for 7260 modem
a new MMGR configuration file has been created. It
contains the same data than configuration #2 excepted
for SSIC link information.
Moreover, because SSIC and HSIC are implementation of
USB link and way of managing them are identical, the
code to use these links are gathered into one tagged
USB. Then all the XML files with hsic and ssic links
are updated to usb link.

Change-Id: Ibf7fbef3eb576220a78ff6573c71458e780b5663
Signed-off-by: Gwenn Bourree <gwenn.bourree@intel.com>