mmgr: optimize PID / VID checking at USB device addition
BZ: 167795

As the USB device opening was done in function is_pid_and_vid
it could be done up to three times (once for 'modem', once for
'flashing' and once for 'core dump' interface types).

Due to the 1 second retry time-out inside the USB Host library,
this was not optimal in case an USB event was detected for a
device that was not the modem (where MMGR does not have access).

By opening the device outside of is_pid_and_vid, the maximum
penalty in this case is only 1 second outside of MMGR's main
loop (as opposed to 3 seconds with the previous code).

Change-Id: I4df3496abaf4264075432759c43a4493d1a7b20d
Signed-off-by: Lionel Ulmer <lionel.ulmer@intel.com>