mmgr: report streamline configuration in case of IPC hang-up
BZ: 169250

Most of the IPC hang-ups are due to faulty streamline step (when
the modem uses the default INBAND configuration).

As it's very easy to have this bad configuration (simply flash a
custom modem without re-doing the streamline step), report the
wake-up configuration in the 'IPC hang up' cases to easily
characterize issues.

Change-Id: I1693449caa12c5b25b474d7fdf3f969aaf320178
Signed-off-by: Lionel Ulmer <lionel.ulmer@intel.com>