mmgr: add new reporting mechanism to crashtool
BZ: 169150

The E_MMGR_NOTIFY_ERROR message is the one used by MMGR to
report event to crashtool. This message is too limitative:

  - only ERROR event can be reported
  - only data0 and data1 fields can be populated
  - no possibility to choose if we want to have the AP or
    the BP logs.

This patch improves the crashtool reporting by adding a new
message: E_MMGR_NOTIFY_TFT_EVENT. With this new message it
is now possible to:

  - report ERROR, STATS or INFO events
  - populate data0 to data5 field
  - chose if we want to have the AP and/or the BP log

The update of the MMGR code to use this new message instead
of the E_MMGR_NOTIFY_ERROR one will be performed in another
patch.

Change-Id: I109660188ebebb53bc546a60f3cb11d343d4cd44
Signed-off-by: Guillaume Lucas <guillaume.lucas@intel.com>