mmgr: replace usage of E_MMGR_NOTIFY_ERROR by E_MMGR_NOTIFY_TFT_EVENT
BZ: 171865

Thanks to the E_MMGR_NOTIFY_TFT_EVENT, the message
E_MMGR_NOTIFY_ERROR is obsoleted and could be replace
by the new one.

This patch updates the code to use the new message
instead of the old one and remove all references to the
old message. Event reported to crashtool are also
updated.

Change-Id: I49b654fc3962105152a67e09a03284c4a4b3ffbd
Signed-off-by: Guillaume Lucas <guillaume.lucas@intel.com>