mmgr: new modem provisioning mechanism
BZ: 172017

The modem provisioning mechanism has been completely reworked.
In POS/ROS and FOTA, the modem updates (fls and tlv) are only
copied in /config/telephony/provisioning. In MOS, MMGR will
check if an update if available before starting the modem. This
check is done before each modem boot to handle modem upgrade in
MOS with miu-app.

An upgrade can be a zip file containing fls and tlv files. In
that case, MMGR needs to extract the good one.
If the upgrade is a fls or a tlv file, this file is copied as it
is.

Change-Id: I380eea9e0fc1d382d1a3f82828232a57fe1840cb
Signed-off-by: Cesar DE OLIVEIRA <cesarx.de.oliveira@intel.com>