mmgr: fix crash of mmgr with user build
BZ: 171570

On user build, there is no libmcdr so the mcdr.hdle is
NULL. The mcdr_dispose() function doesn't check this
variable before to call the dlclose() function. Due to
this missing check, a crash occured in the mmgr.

This patch adds a check before the call to dlclose() in
mcdr_dipose() to avoid the crash.

Change-Id: I189498b58de9f99833af750c3d7557f59aa2d3da
Signed-off-by: Guillaume Lucas <guillaume.lucas@intel.com>