mmgr: call tty_listen_fd of mcdr only if needed
BZ: 172804

If mcdr is not enabled, calling the tty_listen_fd() for the
mcdr will fail with the error "Operation not permitted" as
the mcdr fd is NULL. Due to this error, the user debug
variant can't start the modem.

This patch checks if the mcdr is enabled before to call the
tty_listen_fd() function to avoid the error.

Change-Id: I08fbddffa6d00b2207a016d2b2d366a40e981db4
Signed-off-by: Guillaume Lucas <guillaume.lucas@intel.com>