[Telephony/MMGR]: add new xml config for 7260 cht
BZ: 175770

The Cherrytrail platform does not use the same path
to control the HSIC IPC. The xHCI driver mounts its
PCI device with identifier 14 and not 10 like in the
Moorefield patform. Then a new mmgr configuration file
must be created to support this change.
This change must be implemented in TCS too for the
telephony scalability 7260_CONF_3.

Change-Id: Ib5051ac25d8f5484096e0be4ba197fdc192fe998
Signed-off-by: Gwenn Bourree <gwenn.bourree@intel.com>