mmgr: handle recovery request in WAIT_COLD_ACK state
BZ: 189672

If a client requests a modem recovery when a modem reset is in
progress, this request will be NACKed leading to lost time due
to the client retrying for up to 20 seconds.

Fix the issue by directly ACKing the request (as the modem is
being reset anyway, the client request has been taken into
account).

Change-Id: I778b68e1dfb7a439da524b03cc8dfd61a270efb8
Signed-off-by: Lionel Ulmer <lionel.ulmer@intel.com>