mmgr: Change the main interface PID to select the new CDC mapping.
BZ: 202758

Modify the configuration for MOFD and CHT SSIC.
Du to ANN SSIC issue, only 2 first CDC are available. A new modem firmware
is provided by IMC with a different CDC mapping. Instead of 2ACM+5NCM, it
is now 1ACM+5NCM+1ACM. But modem cannot manage 2 different configurations
with same name.
Thus have to change the PID of the main interface to 1451 to match this
new configuration.

Change-Id: If94a43500d3a8bef27fe4f1654c1b7d2bcd87a82
Signed-off-by: Luc Piguet-Lacroix <lucx.piguet-lacroix@intel.com>