[Telephony/MMGR]: Change SSIC_WA flag definition
BZ: 207479

Because the Scalability V3 overlays mechanism allows
the telephony configuration to load the HW configuration
file and then to be overwritten by other parameters, the
current definition for the SSIC sequences work around flag
is no more relevant. Before the patch it was based on the
telephony configuration to be used. Now, it uses the ipc
control device path and check the presence of the ssic
string in to be set.

Change-Id: I9ce7e7fd3f44c13133cde603ce75557f95b77476
Signed-off-by: Gwenn Bourree <gwenn.bourree@intel.com>