mmgr:Fix compilation issues with Imin branch
BZ: 214617

The LOG(X) macros are removed by google sometime back.
We could still use it because of an AOSP patch. But
this patch is no more available with imin.

The usage of LOCAL_IMPORT_C_INCLUDE_DIRS_FROM_SHARED_LIBRARIES
is not yet supported and is removed from .mk files.

mizip changed with imin_legacy and the usage is changed in mmgr too.
Wrong usage of pthread as LDFLAGS removed.

Other changes are done by the indent script.

Change-Id: I31456d4380e38ad521e0a38b1e5291b40adaa8f0
Signed-off-by: Arun Ravindran <arun.ravindran@intel.com>
Signed-off-by: CristianX Mateescu <cristianx.mateescu@intel.com>