MMGR: MDM cold reset sequence fix
BZ: 203982

 The MDM cold reset sequence should start by waiting for ack
from subscribed clients before the reset is issued. In case
no clients are connected then the reset will proceed.
  Replaced the static variable 'wait_operation' with the reset
escalation states in order to be able to reset it at the end of the
escalation process. The variable was remaining stuck for
the case when no clients were connected and the next cold reset
started without clients acknowledge.
  Also replaced the E_OPERATION_UNKNOWN with E_OPERATION_NONE and
removed the unused state(E_OPERATION_SKIP) from reset
escalation.

Change-Id: Ia7942d982800b932a214774561998470c05ee556
Signed-off-by: Alexandru <ioan.alexandrux.ratiu@intel.com>
Signed-off-by: Florin Dracman <florin.zoreanux.dracman@intel.com>