mmgr: do not send COLD_RESET intent after TLV application
BZ: 219895

As the modem reset following the TLV application is part of the
normal phone operation, do not send the COLD_RESET intent as it
may confuse the tester.

The filtering is done by setting the recovery operation to
FORCE_NO_COUNT. This has the additional benefit of not counting
the cold reset linked to the TLV application in the reset
escalation process.

Change-Id: I42d00dfa63496a042d10b7ad306c6c9ed9b70141
Signed-off-by: Lionel Ulmer <lionel.ulmer@intel.com>
Signed-off-by: Riviere, Nathan <nathanx.riviere@intel.com>