mmgr: fix handling of flashing file descriptor
BZ: 219894

The file descriptor used to communicate with the flashing thread
was only added to the poll set if a .fls file was present in the
configuration directory. This meant that the first flash after a
blankphone would work ... but would not be detected as such due
to the fact that the file descriptor was not polled in the main
MMGR thread.

Fix the issue by always put the flashing file descriptor in the
poll set as it is now created unconditionnaly (i.e. both in
flashbased and flashless configurations).

Change-Id: Ic09c3c22d507cca6a33bb26e9b6b5af6ac69fa39
Signed-off-by: Lionel Ulmer <lionel.ulmer@intel.com>
Signed-off-by: Riviere, Nathan <nathanx.riviere@intel.com>