mmgr: during IPC hangup send modem down event before ipc cleanup
During M.2 modem core dump sometimes MMGR is not able to capture
the coredump because of a client requesting modem recovery
before MMGR starts to handle the event. This is caused by the
fact that MMGR is doing first the cleanup for the IPC event
(closing the tty's) and only after informing the clients of
modem down event.

In order to fix this the MMGR should inform clients immediately
of the modem down event, and only after handle the more time
consuming action of cleanup ipc.

Change-Id: Icaeb7db506b6eb21eb8d93de74d72cc5dcabf0e0
Tracked-On: https://jira01.devtools.intel.com/browse/IMINAN-3930
Signed-off-by: George Sandu <georgex.sandu@intel.com>
Reviewed-on: https://android.intel.com:443/283137