mmgr: IPC configuration and SPI support
When an interface appears, some configuration needs to be done.
Especially for the flashing and baseband interfaces.

HSI interface was configured in MUP and SPI configuration was not
configured.

This patch tries to bring together all those operations in a new
module called link. Power management, link control and
configuration are managed by this module.

XML files have been updated to add the needed link parameters.

Change-Id: I13adb5de2762152817f6b7f358b2aa4414e1ff98
Signed-off-by: Cesar DE OLIVEIRA <cesar.de.oliveira@intel.com>
Tracked-On: https://jira01.devtools.intel.com/browse/IMINAN-5297
Signed-off-by: Nicolas LOUIS <nicolasx.louis@intel.com>
Reviewed-on: https://android.intel.com:443/287425