mmgr: add pwr_on_ctrl and usb_hub_ctrl in mmgr mdm_ctrl section
In order to support different hardware control methods for the
modem power_on signal a new entry was added to the MMGR config
file to specify the type of power_on control (PMIC/GPIO).
In order to enable/disable the reseting of the USB to HSIC hub
present on some platforms a new entry was added to mmgr config
to inform the MCD of the need to control this signal.

Change-Id: Ie9ce64497b8f858ba862f777216f3404d3a02d23
Tracked-On: https://jira01.devtools.intel.com/browse/IMINAN-7979
Signed-off-by: George Sandu <georgex.sandu@intel.com>
Reviewed-on: https://android.intel.com:443/299732