mmgr: data partition encryption
MMGR uses persistent properties to detect a modification of
its configuration. Those properties are not accessible while the
data partition is encrypted leading to a false factory reset
detection. As a consequence, dynamic NVM parameters are deleted
and TLV files are reapplied.

To fix this issue, MMGR does not check those properties when the
data partition is encrypted and the modem upgrade is disabled.
Once the partition is decrypted, MMGR will be restarted and act
normally.

This patch fixes also some coding style issues because some
contributors have forgotten to run the indent script...

Change-Id: Idef5790e939fb5d5a6fcec97b11d2db21d56910a
Tracked-On: https://jira01.devtools.intel.com/browse/IMINAN-10633
Signed-off-by: Cesar DE OLIVEIRA <cesar.de.oliveira@intel.com>
Reviewed-on: https://android.intel.com:443/306916