mmgr: kill NVM server in case of MUX failure
Due to an issue in the NVM server, it can block during modem
shutdown with a GSM TTY still open. This will prevent the MMGR
to mux the modem TTY at next modem start leading to a platform
shutdown.

Fix the issue by killing NVM server when MMGR cannot attach the
new line discipline.

This is a work-around waiting for the issue to be fixed in NVM
server.

Change-Id: I047331826b2f149f603d7855c81506cf8fc908ce
Tracked-On: https://jira01.devtools.intel.com/browse/IMINAN-12084
Signed-off-by: Lionel Ulmer <lionel.ulmer@intel.com>
Reviewed-on: https://android.intel.com:443/311263